// src/api/generated/model/suggestedRuleOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * LeadRadar API
 * OpenAPI spec version: 0.1.0
 */
import type { RuleAction } from './ruleAction';
import type { RuleCondition } from './ruleCondition';
import type { RuleKind } from './ruleKind';

export interface SuggestedRuleOut {
  name: string;
  kind: RuleKind;
  condition: RuleCondition;
  action: RuleAction;
  cap_value: number | null;
}